add example of stateful hypothesis testing
Notice that the state machine being tested (`TransmissionSystem`) is
not itself a subclass of hypothesis'
`RuleBasedStateMachine` (`RBSM`). Instead, a separate class
`AutoFSMTest` is the `RBSM` subclass. It creates an instance of and
then exercises the `TransmissionSystem`. `AutoFSMTest` also happens to
be a state machine, but it is used to test a different state machine.